Check for zero vocab, reduce epochs and fix inference step
Makes some small fixes. Ensures rank_zero_info is available for outputing the
inference step at the end. Reduces the number of training epochs from 10 to 2
for use as an example. Disables checkpointing. Adds an assertion check for a
non-zero vocabulary size to catch cases where the training data isn't read in
correctly.